+++ .jules/tidy.md
@@ -145,3 +145,11 @@ out` として共通化することで保守性が向上した。
 ## 2026-03-03 - [Tidy: リファクタリング - 設定画面ダイアログのUI共通化]
 **学び:** 設定画面（赤ちゃん追加、編集、アプリ情報、メンバーロール変更など）の各ダイアログにおいて、`EditDialogBase` 共通コンポーネントが活用されず、個別に `Dialog`, `DialogContent`, `DialogHeader` などが呼び出されていました。しかし、単なる置換では、固有のレイアウト制約（`max-h-[90vh]` など）や `DialogFooter` のレスポンシブな挙動、フォーカス管理 (`onOpenAutoFocus`) が失われ、UIの崩れやアクセシビリティの低下を招くリスク（アンチパターン）があることが分かりました。
 **アクション:** `EditDialogBase` に対して、外部から柔軟にスタイルや挙動を注入できる `contentClassName`, `dialogClassName`, `onOpenAutoFocus` などのPropsを追加した上で、設定画面の各ダイアログをリファクタリングしました。今後、共通コンポーネントを適用する際は、元のUIの特性（レスポンシブ性やスクロール領域、フォーカス挙動）を損なわないよう、必要に応じてベースコンポーネントのインターフェース拡張を慎重に行うことを徹底します。
+
+## 2024-03-08 - useEffectの抽出と共通化
+
+**学び:**
+`useEffect`を用いてマウント状態（`mounted`や`isMounted`）を遅延評価するロジック（ハイドレーション時の警告やエラー回避のための手法など）が、複数のコンポーネント（`app/page.tsx`, `components/ui/splash-screen.tsx`, `components/ui/tips-card.tsx`）で直接実装されていました。これを個別に書くと、実装のばらつきやコードの肥大化、保守性の低下につながります。
+
+**アクション:**
+これらをカスタムフック `useMounted` に抽出し、共通化しました。これにより、各コンポーネントのコードが簡潔になり、意図（「マウント完了を待機する」）が明確に伝わるようになります。今後は、他のコンポーネントでも同様の遅延マウント判定が必要になった場合、この `useMounted` フックを再利用するようにします。
